+++ adl.c
@@ -121,7 +121,6 @@ static	ADL_OVERDRIVE6_POWERCONTROL_SET		ADL_Overdrive6_PowerControl_Set;
 #endif
 static int iNumberAdapters;
 static LPAdapterInfo lpInfo = NULL;
-static LPADLDisplayInfo lpAdlDisplayInfo = NULL;
 
 int set_fanspeed(int gpu, int iFanSpeed);
 static float __gpu_temp(struct gpu_adl *ga);
@@ -326,7 +325,6 @@ void init_adl(int nDevs)
 	struct gpu_adapters adapters[MAX_GPUDEVICES], vadapters[MAX_GPUDEVICES];
 	bool devs_match = true;
 	ADLBiosInfo BiosInfo;
-	int iNumDisplays;
 
 	applog(LOG_INFO, "Number of ADL devices: %d", nDevs);
 
@@ -361,7 +359,7 @@ void init_adl(int nDevs)
 		return;
 	}
 
-	applog(LOG_INFO, "Found %d ADL adapters", iNumberAdapters);
+	applog(LOG_INFO, "Found %d logical ADL adapters", iNumberAdapters);
 
 	/* Iterate over iNumberAdapters and find the lpAdapterID of real devices */
 	for (i = 0; i < iNumberAdapters; i++) {