+++ algorithm.h
@@ -2,6 +2,7 @@
 #define ALGORITHM_H
 
 #include <inttypes.h>
+#include <stdbool.h>
 
 struct work;
 
@@ -39,9 +40,12 @@ typedef struct _algorithm_t {
 } algorithm_t;
 
 /* Set default parameters based on name. */
-void set_algorithm(algorithm_t** algo, const char* name);
+void set_algorithm(algorithm_t* algo, const char* name);
 
 /* Set to specific N factor. */
 void set_algorithm_nfactor(algorithm_t* algo, const uint8_t nfactor);
+
+/* Compare two algorithm parameters */
+bool cmp_algorithm(algorithm_t* algo1, algorithm_t* algo2);
 
 #endif /* ALGORITHM_H */